Avoid ClassCastException for the same class when used with Flink
https://ci.apache.org/projects/flink/flink-docs-release-1.2/monitoring/debugging_classloading.html#x-cannot-be-cast-to-x-exceptions

It is documented that when Avro is used with Flink, a `ClassCastException`
is thrown, due to a caching issue. The issue occurs, then the confluent
kafka client is configured with `specific.avro.reader= true`, and the
de-serialization is performed with the `SpecificDatumReader`, instead of
the generic reader.
Based on Flink documentation, it is recommended to pass a new
`SpecificData` to `SpecificDatumReader` each time.

There is JIRA issue in Flink discussing this bug:
https://issues.apache.org/jira/browse/FLINK-5633

This hotfix is backported PR fixing this problem and
waiting to be merged here:
https://github.com/confluentinc/schema-registry/pull/509